alx_travel_app/alx_travel_app/requirement.txt
asgiref==3.9.1
django==5.2.6
django-countries==7.6.1
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.13
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
pyjwt==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
sqlparse==0.5.3
typing-extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0